# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/admin/generate_video_stream/": {
            "post": {
                "description": "api to get the list of all the videos uploaded by the logged user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "VideoManagement"
                ],
                "summary": "Generate video stream",
                "parameters": [
                    {
                        "description": "Video IDs",
                        "name": "video_ids",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/admin_views.VideoStreamReqStruct"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/my_modules.ResponseFormat"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/my_modules.ResponseFormat"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/my_modules.ResponseFormat"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/my_modules.ResponseFormat"
                        }
                    }
                }
            }
        },
        "/admin/get_stream_key/": {
            "post": {
                "description": "api to get video decryption key for hls stream",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "VideoManagement"
                ],
                "summary": "get video decode key",
                "parameters": [
                    {
                        "description": "Video ID",
                        "name": "video_id",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/admin_views.VideoStreamKeyReqStruct"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/my_modules.ResponseFormat"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/my_modules.ResponseFormat"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/my_modules.ResponseFormat"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/my_modules.ResponseFormat"
                        }
                    }
                }
            }
        },
        "/admin/playlist/": {
            "post": {
                "description": "api to create new empty playlist",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Playlist"
                ],
                "summary": "Create new playlist",
                "parameters": [
                    {
                        "description": "New Playlist",
                        "name": "new_playlist_data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/database.VideoPlayListModal"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/my_modules.ResponseFormat"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/my_modules.ResponseFormat"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/my_modules.ResponseFormat"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/my_modules.ResponseFormat"
                        }
                    }
                }
            }
        },
        "/admin/upload_list/": {
            "get": {
                "description": "api to get the list of all the videos uploaded by the logged user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "VideoManagement"
                ],
                "summary": "get all uploaded videos",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/my_modules.ResponseFormat"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/my_modules.ResponseFormat"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/my_modules.ResponseFormat"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/my_modules.ResponseFormat"
                        }
                    }
                }
            }
        },
        "/admin/upload_streaming_video/": {
            "post": {
                "description": "api to upload video content for multiple adaptive bit rate streaming",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "VideoManagement"
                ],
                "summary": "video upload",
                "parameters": [
                    {
                        "type": "file",
                        "description": "Video file",
                        "name": "video_file",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "Preview image file",
                        "name": "preview_image_file",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "created_by",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "description",
                        "in": "formData"
                    },
                    {
                        "type": "boolean",
                        "name": "is_live",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "name": "title",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/my_modules.ResponseFormat"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/my_modules.ResponseFormat"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/my_modules.ResponseFormat"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/my_modules.ResponseFormat"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "admin_views.VideoStreamKeyReqStruct": {
            "type": "object",
            "required": [
                "video_id"
            ],
            "properties": {
                "video_id": {
                    "type": "string"
                }
            }
        },
        "admin_views.VideoStreamReqStruct": {
            "type": "object",
            "required": [
                "video_ids"
            ],
            "properties": {
                "video_ids": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "database.VideoPlayListModal": {
            "type": "object",
            "required": [
                "is_live",
                "price",
                "title",
                "videos_ids"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "is_live": {
                    "type": "boolean"
                },
                "price": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "videos_ids": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "my_modules.ResponseFormat": {
            "type": "object",
            "required": [
                "data",
                "msg",
                "status"
            ],
            "properties": {
                "data": {},
                "msg": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
    }
}